+++ docs/CHANGES.rst
@@ -7,6 +7,8 @@ Changelog
 - Added do_rendering field in IMergeTemplatesRowSchema schema. If selected, the subtemplate is rendered first
   and the path is the value in context dict. Else the subtemplate object is the value in context dict.
   [sgeulette]
+- Return generation context from rendering methods to use it in tests
+  [sgeulette]
 
 1.0.5 (2017-03-10)
 ------------------